Refactor: SIZE를 정의하는 파일명 변경 및 적용, COLOR_SET 적용
