+++ src/components/list/ListContent.jsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import RecipientCardList from './RecipientCardList';
 import { bold24 } from '../../styles/fontStyle';
-import { DISPLAY_SIZE } from '../../constants/DISPLAY_SIZE';
+import { DISPLAY_SIZE } from '../../constants/SIZE_SET';
 
 function ListContent({ title, recipients }) {
   return (
