+++ src/components/list/RecipientCard.jsx
@@ -4,7 +4,7 @@ import styled, { css } from 'styled-components';
 import WrittenByIcons from '../post/subheader/WrittenByIcons';
 import { getAllMessages } from '../../api/GetApi';
 import { bold18, bold24, regular14, regular16 } from '../../styles/fontStyle';
-import { DISPLAY_SIZE } from '../../constants/DISPLAY_SIZE';
+import { DISPLAY_SIZE } from '../../constants/SIZE_SET';
 
 function RecipientCard({ recipient }) {
   const backgroundColor = recipient.backgroundColor || 'beige';
