+++ src/components/list/RecipientCardList.jsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 import styled from 'styled-components';
 import RecipientCard from './RecipientCard';
-import { DISPLAY_SIZE } from '../../constants/DISPLAY_SIZE';
+import { DISPLAY_SIZE } from '../../constants/SIZE_SET';
 
 function RecipientList({ recipients }) {
   const containerRef = useRef(null);
