+++ src/components/main/MainSection.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import MovePageButton from '../common/Buttons/movePageButton';
 import { bold14, bold24, regular18 } from '../../styles/fontStyle';
-import { DISPLAY_SIZE } from '../../constants/DISPLAY_SIZE';
+import { DISPLAY_SIZE } from '../../constants/SIZE_SET';
 
 const StyledSection = styled.section`
   display: flex;
