+++ src/components/post/Card.jsx
@@ -9,14 +9,9 @@ import {
   regular18,
   regular20,
 } from '../../styles/fontStyle';
-import { DISPLAY_SIZE } from '../../constants/DISPLAY_SIZE';
+import { DISPLAY_SIZE } from '../../constants/SIZE_SET';
+import { USER_STATE } from '../../constants/COLOR_SET';
 
-const userStateColors = {
-  가족: { background: 'var(--green100)', color: 'var(--green500)' },
-  동료: { background: 'var(--purple100)', color: 'var(--purple600)' },
-  지인: { background: 'var(--orange100)', color: 'var(--orange500)' },
-  친구: { background: 'var(--blue100)', color: 'var(--blue500)' },
-};
 export const CardContentWrapper = styled.div`
   position: relative;
   max-width: 38.4rem;
@@ -91,11 +86,9 @@ const UserState = styled.div`
 
   border-radius: 0.4rem;
   background: ${({ $state }) =>
-    userStateColors[$state]
-      ? userStateColors[$state].background
-      : 'defaultColor'};
+    USER_STATE[$state] ? USER_STATE[$state].background : 'defaultColor'};
   color: ${({ $state }) =>
-    userStateColors[$state] ? userStateColors[$state].color : 'defaultColor'};
+    USER_STATE[$state] ? USER_STATE[$state].color : 'defaultColor'};
 
   ${regular14}
 `;
