+++ src/components/post/postmake/PostInputSection.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { bold24, regular12, regular16 } from '../../../styles/fontStyle';
-import { DISPLAY_SIZE } from '../../../constants/DISPLAY_SIZE';
+import { DISPLAY_SIZE } from '../../../constants/SIZE_SET';
 
 function PostInputSection({ user, handleInput, handleBlur, inputError }) {
   return (
